fix: change lang button position
